+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased 
+## Unreleased [0.3.0-SNAPSHOT]
 ### Added
+- Relational DAO now adds functions to query for indexed fields
+- API @GetBy
+- Generate API Query Models
 
 ### Changed